chore: clean up & renaming

+++ projects/frontend/es-core/src/utils/math/Transform4x4f.cpp
@@ -139,8 +139,8 @@ Transform4x4f& Transform4x4f::rotate(const float _angle, const Vector3f& _axis)
 {
 	float*       tm      = (float*)this;
 	const float* av      = (float*)&_axis;
-	const float  s       = Math::sinf(-_angle);
-	const float  c       = Math::cosf(-_angle);
+	const float  s       = Math::sin(-_angle);
+	const float  c       = Math::cos(-_angle);
 	const float  t       = 1 - c;
 	const float  x       = av[0];
 	const float  y       = av[1];
@@ -187,8 +187,8 @@ Transform4x4f& Transform4x4f::rotate(const float _angle, const Vector3f& _axis)
 Transform4x4f& Transform4x4f::rotateX(const float _angle)
 {
 	float*      tm      = (float*)this;
-	const float s       = Math::sinf(-_angle);
-	const float c       = Math::cosf(-_angle);
+	const float s       = Math::sin(-_angle);
+	const float c       = Math::cos(-_angle);
 	const float temp[6] = { tm[ 1] *  c + tm[ 2] * s,
 		                    tm[ 1] * -s + tm[ 2] * c,
 		                    tm[ 5] *  c + tm[ 6] * s,
@@ -210,8 +210,8 @@ Transform4x4f& Transform4x4f::rotateX(const float _angle)
 Transform4x4f& Transform4x4f::rotateY(const float _angle)
 {
 	float*      tm      = (float*)this;
-	const float s       = Math::sinf(-_angle);
-	const float c       = Math::cosf(-_angle);
+	const float s       = Math::sin(-_angle);
+	const float c       = Math::cos(-_angle);
 	const float temp[6] = { tm[ 0] * c + tm[ 2] * -s,
 		                    tm[ 0] * s + tm[ 2] *  c,
 		                    tm[ 4] * c + tm[ 6] * -s,
@@ -233,8 +233,8 @@ Transform4x4f& Transform4x4f::rotateY(const float _angle)
 Transform4x4f& Transform4x4f::rotateZ(const float _angle)
 {
 	float*      tm      = (float*)this;
-	const float s       = Math::sinf(-_angle);
-	const float c       = Math::cosf(-_angle);
+	const float s       = Math::sin(-_angle);
+	const float c       = Math::cos(-_angle);
 	const float temp[6] = { tm[ 0] *  c + tm[ 1] * s,
 		                    tm[ 0] * -s + tm[ 1] * c,
 		                    tm[ 4] *  c + tm[ 5] * s,
